+++ NEWS.md
@@ -1,3 +1,9 @@
-# pivoteasy 0.0.0.9000
+# pivoteasy release news
 
-* Added a `NEWS.md` file to track changes to the package.
+# pivoteasy 0.9.9.9000
+
+* 2023-06-15
+
+* First release
+    * Add `pivot()` and related functions: `add_group_sub()`, `validate_col()` and `split_force()`
+    * Add util functions: `na_row_omit()`, `has_col()`, `na_col_omit()`, `na_count()`, `extract_col()`, and `na2empty()`